backend/requirements.txt
langchain
requests
python-dotenv
fastapi
uvicorn
